Only link leveldb with applications that need it
Leveldb was being linked with all applications, regardless
if they needed it or not.

There is still an issue where librados.so.2 somehow still
has a dependency on libleveldb.

This is how I test:
./autogen.sh
./configure
cd src
V=1 make rados
ldd .src/rados

I even created an hello world application and linked it
to librados, and the executable also showed a dependency
on libleveldb.